Refactor 'sendMail' to a general helper
