+++ backend/src/middleware/email/emailMiddleware.js
@@ -1,46 +1,10 @@
-import CONFIG from '../../config'
-import nodemailer from 'nodemailer'
-import { htmlToText } from 'nodemailer-html-to-text'
+import { sendMail } from '../helpers/email/sendMail'
 import {
   signupTemplate,
   resetPasswordTemplate,
   wrongAccountTemplate,
   emailVerificationTemplate,
-} from './templateBuilder'
-
-const hasEmailConfig = CONFIG.SMTP_HOST && CONFIG.SMTP_PORT
-const hasAuthData = CONFIG.SMTP_USERNAME && CONFIG.SMTP_PASSWORD
-
-let sendMail = () => {}
-if (!hasEmailConfig) {
-  if (!CONFIG.TEST) {
-    // eslint-disable-next-line no-console
-    console.log('Warning: Email middleware will not try to send mails.')
-  }
-} else {
-  sendMail = async (templateArgs) => {
-    const transporter = nodemailer.createTransport({
-      host: CONFIG.SMTP_HOST,
-      port: CONFIG.SMTP_PORT,
-      ignoreTLS: CONFIG.SMTP_IGNORE_TLS,
-      secure: CONFIG.SMTP_SECURE, // true for 465, false for other ports
-      auth: hasAuthData && {
-        user: CONFIG.SMTP_USERNAME,
-        pass: CONFIG.SMTP_PASSWORD,
-      },
-    })
-
-    transporter.use(
-      'compile',
-      htmlToText({
-        ignoreImage: true,
-        wordwrap: false,
-      }),
-    )
-
-    await transporter.sendMail(templateArgs)
-  }
-}
+} from '../helpers/email/templateBuilder'
 
 const sendSignupMail = async (resolve, root, args, context, resolveInfo) => {
   const { inviteCode } = args
